refactor(skribenten-backend): flytt getAlltidValgbareVedlegg til handler
Konverterer BrevmalService.getAlltidValgbareVedlegg til
HentAlltidValgbareVedleggHandler etter samme mønster som de andre
brevredigering-handlerne (TransactionHandler). Ruten i SakBrev.kt
bruker nå respondOutcome i stedet for manuell null-sjekk.
