+++ skribenten-backend/src/main/kotlin/no/nav/pensjon/brev/skribenten/fagsystem/BrevmalService.kt
@@ -7,7 +7,6 @@ import no.nav.pensjon.brev.api.model.TemplateDescription
 import no.nav.pensjon.brev.api.model.maler.Brevkode
 import no.nav.pensjon.brev.skribenten.brevbaker.BrevbakerService
 import no.nav.pensjon.brev.skribenten.brevredigering.domain.Brevredigering
-import no.nav.pensjon.brev.skribenten.brevredigering.domain.BrevredigeringEntity
 import no.nav.pensjon.brev.skribenten.common.GeneriskRedigerbarBrevdata
 import no.nav.pensjon.brev.skribenten.fagsystem.pesys.BrevdataDto
 import no.nav.pensjon.brev.skribenten.fagsystem.pesys.BrevdataResponse
@@ -22,17 +21,9 @@ import no.nav.pensjon.brevbaker.api.model.LetterMarkup
 import no.nav.pensjon.brevbaker.api.model.LetterMarkupWithDataUsage
 import no.nav.pensjon.brevbaker.api.model.RedigerbareVedleggTitler
 import no.nav.pensjon.brevbaker.api.model.TemplateModelSpecification
-import org.jetbrains.exposed.v1.jdbc.transactions.transaction
 import org.slf4j.LoggerFactory
 
 private val ekskluderteBrev = hashSetOf("PE_IY_05_301", "PE_BA_01_108", "PE_GP_01_010", "PE_AP_04_922", "PE_IY_03_169")
-
-data class ValgbartVedlegg(
-    val kode: String,
-    val visningstekst: String,
-    val spraak: Set<LanguageCode>,
-    val tilgjengeligForSpraak: Boolean,
-)
 
 class BrevmalService(
     private val brevbakerService: BrevbakerService,
@@ -96,20 +87,6 @@ class BrevmalService(
 
     suspend fun getModelSpecification(brevkode: Brevkode.Redigerbart): TemplateModelSpecification? =
         brevbakerService.getModelSpecification(brevkode)
-
-    suspend fun getAlltidValgbareVedlegg(brevId: BrevId): List<ValgbartVedlegg>? {
-        val spraakIBrevet = transaction {
-            BrevredigeringEntity.findById(brevId)?.spraak
-        } ?: return null
-        return brevbakerService.getAlltidValgbareVedlegg(brevId).map {
-            ValgbartVedlegg(
-                kode = it.kode,
-                visningstekst = it.visningstekst,
-                spraak = it.spraak,
-                tilgjengeligForSpraak = it.spraak.contains(spraakIBrevet),
-            )
-        }.sortedBy { it.visningstekst }
-    }
 
     suspend fun getTemplates(): List<TemplateDescription.Redigerbar>? =
         brevbakerService.getTemplates()
